+++ config/routes.rb
@@ -87,7 +87,9 @@ Rails.application.routes.draw do
     # for admin
     resources :trainings
     resources :credits
-    resources :categories, only: [:index]
+    resources :categories
+    resources :event_themes
+    resources :age_ranges
     resources :statistics, only: [:index]
     resources :custom_assets, only: [:show, :create, :update]
     resources :tags