+++ db/migrate/20160630083556_add_slug_to_age_range.rb
@@ -1,0 +1,6 @@
+class AddSlugToAgeRange < ActiveRecord::Migration
+  def change
+    add_column :age_ranges, :slug, :string
+    add_index :age_ranges, :slug, unique: true
+  end
+end